Revert "Auth0 Login and uploads page"
