--- .env.example
@@ -1,4 +1,0 @@
-VITE_AUTH0_DOMAIN=dev-hwbeuz3j71v25bdy.us.auth0.com
-VITE_AUTH0_CLIENT_ID=hG195SeUNV6OzBfb4xP83MLWSp5vXu5v
-VITE_AUTH0_AUDIENCE=https://applens-api
-VITE_API_BASE_URL=http://localhost:8080
