+++ .gitignore
@@ -23,4 +23,3 @@ dist-ssr
 *.sln
 *.sw?
 .claude
-.env
