+++ package.json
@@ -13,13 +13,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@auth0/auth0-react": "^2.15.1",
-    "@js-temporal/polyfill": "^0.5.1",
     "@tailwindcss/vite": "^4.1.18",
-    "lucide-react": "^1.0.1",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
-    "react-router-dom": "^6.26.2",
     "tailwindcss": "^4.1.18"
   },
   "devDependencies": {
