+++ src/index.css
@@ -38,156 +38,3 @@ body {
   -webkit-mask-image: linear-gradient(to left, black 0%, transparent 96%);
   mask-image: linear-gradient(to left, black 0%, transparent 96%);
 }
-
-section.dashboard {
-  /* padding-top: 50px; */
-  padding-top: 30px;
-}
-
-.column-parent {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  justify-content: space-between;
-}
-
-.column-parent .left-column, .column-parent .right-column {
-  display: inline-block;
-  padding: 10px 20px;
-  padding-top: 20px;
-}
-
-.left-column {
-  flex: 1;
-}
-
-.right-column {
-  flex: 2;
-}
-
-
-.risk-score-container{
-  padding: 30px;
-}
-
-.risk-score-container {
-  width: 350px;
-  padding: 75px 50px;
-  border-radius: 50%;
-  margin-bottom: 50px;
-  border: 5px solid white;
-}
-
-.risk-score-container.green {
-  background-color: rgb(27, 226, 27);
-}
-
-.risk-score-container.yellow {
-  background-color: rgb(226, 230, 5);
-}
-
-.risk-score-container.red {
-  background-color: rgb(235, 72, 32);
-}
-
-.risk-score-container .risk-score {
-  background-color: black;
-  width: 180px;
-  margin: 0 auto;
-  padding: 40px;
-  border-radius: 50%;
-  text-align: center;
-  border: 5px solid white;
-}
-
-.risk-score h2 {
-  font-size: 17px;
-}
-
-.risk-score p {
-  font-size: 50px;
-}
-
-.vulnerability-list {
-  list-style: decimal inside;
-  border: 1px solid white;
-  padding: 15px;
-  min-height: 342px;
-  width: 430px; 
-}
-
-.vulnerability-list .vulnerability-row {
-  border: 1px solid white;
-  cursor: pointer;
-  padding: 4px;
-}
-
-.vulnerability-list .green-badge {
-  padding: 8px;
-  border-radius: 20px;
-  background-color: rgb(27, 226, 27);
-  display: inline-block;
-  height: 5px;
-  margin-left: 5px;
-}
-
-.vulnerability-list .yellow-badge {
-  padding: 8px;
-  border-radius: 20px;
-  background-color: rgb(226, 230, 5);
-  display: inline-block;
-  height: 5px;
-  margin-left: 5px;
-}
-
-.vulnerability-list .red-badge {
-  padding: 8px;
-  border-radius: 20px;
-  background-color: rgb(235, 72, 32);
-  display: inline-block;
-  height: 5px;
-  margin-left: 5px
-}
- .vulnerability-list .icon {
-   float: right;
- }
-
-.right-column p {
-  text-align: right;
-  padding: 0 10px 10px 10px;
-  margin-top: 0;
-  font-size: 25px;
-  padding-bottom: 20px;
-}
-
-.right-column img {
-  width: 100%;
-}
-
-.timeline {
-  width: 100%;
-  border: 1px solid gray;
-  margin-top: 50px;
-  min-height: 200px;
-  padding: 15px;
-}
-
-.timeline p {
-  text-align: center;
-  margin-top: 40px;
-  font-size: 15px;
-}
-
-.timeline-bar {
-  border-bottom: 1px solid white;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.timeline-bar span {
-  height: 20px;
-  border-left: 1px solid gray;
-  position: relative;
-  bottom: -10px;
-}
