+++ src/main.tsx
@@ -1,43 +1,10 @@
-import { StrictMode } from "react";
-import type { ReactNode } from "react";
-import { createRoot } from "react-dom/client";
-import { BrowserRouter, useNavigate } from "react-router-dom";
-import { Auth0Provider } from "@auth0/auth0-react";
-import type { AppState } from "@auth0/auth0-react";
-import "./index.css";
-import App from "./App";
+import { StrictMode } from 'react'
+import { createRoot } from 'react-dom/client'
+import './index.css'
+import App from './App.tsx'
 
-export function Auth0ProviderWithNavigate({ children }: { children: ReactNode }) {
-  const navigate = useNavigate();
-
-  const onRedirectCallback = (appState?: AppState) => {
-    navigate(appState?.returnTo || "/onboarding/import-terraform", { replace: true });
-  };
-
-  return (
-    <Auth0Provider
-      domain={import.meta.env.VITE_AUTH0_DOMAIN}
-      clientId={import.meta.env.VITE_AUTH0_CLIENT_ID}
-      authorizationParams={{
-        redirect_uri: window.location.origin,
-        audience: import.meta.env.VITE_AUTH0_AUDIENCE,
-        scope: "openid profile email",
-      }}
-      onRedirectCallback={onRedirectCallback}
-      cacheLocation="localstorage"
-      useRefreshTokens={true}
-    >
-      {children}
-    </Auth0Provider>
-  );
-}
-
-createRoot(document.getElementById("root")!).render(
+createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <BrowserRouter basename={import.meta.env.BASE_URL}>
-      <Auth0ProviderWithNavigate>
-        <App />
-      </Auth0ProviderWithNavigate>
-    </BrowserRouter>
-  </StrictMode>
-);
+    <App />
+  </StrictMode>,
+)
